backend/pyproject.toml
[project]
name = "nova-sonic-healthcare-backend"
version = "0.1.0"
description = "Nova Sonic Healthcare Call Center - BidiAgent Backend"
requires-python = ">=3.12"
dependencies = [
    # Strands Agents SDK for BidiAgent (includes strands.experimental.bidi)
    "strands-agents>=0.1.0",
    # PyAudio for strands.experimental.bidi imports (requires portaudio system lib)
    "pyaudio>=0.2.14",
    # prompt_toolkit for strands.experimental.bidi text IO
    "prompt_toolkit>=3.0.0",
    # AWS SDK
    "boto3>=1.34.0",
    # AWS SDK for Bedrock Runtime bidirectional streaming (required by strands BidiNovaSonicModel)
    "aws-sdk-bedrock-runtime>=0.3.0",
    # FastAPI and WebSocket server for AgentCore Runtime
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=14.0",
    # Logging
    "python-json-logger>=2.0.0",
]

[tool.pylint.main]
ignore = [".venv"]
